+++ core/manifest/dbkeywords.go
@@ -1,79 +1,84 @@
 package manifest
 
 import (
+	"embed"
+
 	"github.com/opensvc/om3/core/keywords"
 	"github.com/opensvc/om3/util/converters"
 )
+
+//go:embed text
+var fs embed.FS
 
 var (
 	KWBlockingPostProvision = keywords.Keyword{
 		Option:   "blocking_post_provision",
 		Attr:     "BlockingPostProvision",
 		Scopable: true,
-		Text:     "A command or script to execute after the resource :c-action:`provision` action. Errors interrupt the action.",
+		Text:     keywords.NewText(fs, "text/kw/blocking_post_provision"),
 	}
 
 	KWBlockingPostRun = keywords.Keyword{
 		Option:   "blocking_post_run",
 		Attr:     "BlockingPostRun",
 		Scopable: true,
-		Text:     "A command or script to execute after the resource :c-action:`run` action. Errors interrupt the action.",
+		Text:     keywords.NewText(fs, "text/kw/blocking_post_run"),
 	}
 
 	KWBlockingPostStart = keywords.Keyword{
 		Option:   "blocking_post_start",
 		Attr:     "BlockingPostStart",
 		Scopable: true,
-		Text:     "A command or script to execute after the resource :c-action:`start` action. Errors interrupt the action.",
+		Text:     keywords.NewText(fs, "text/kw/blocking_post_start"),
 	}
 
 	KWBlockingPostStop = keywords.Keyword{
 		Option:   "blocking_post_stop",
 		Attr:     "BlockingPostStop",
 		Scopable: true,
-		Text:     "A command or script to execute after the resource :c-action:`stop` action. Errors interrupt the action.",
+		Text:     keywords.NewText(fs, "text/kw/blocking_post_stop"),
 	}
 
 	KWBlockingPostUnprovision = keywords.Keyword{
 		Option:   "blocking_post_unprovision",
 		Attr:     "BlockingPostUnprovision",
 		Scopable: true,
-		Text:     "A command or script to execute after the resource :c-action:`unprovision` action. Errors interrupt the action.",
+		Text:     keywords.NewText(fs, "text/kw/blocking_post_unprovision"),
 	}
 
 	KWBlockingPreProvision = keywords.Keyword{
 		Option:   "blocking_pre_provision",
 		Attr:     "BlockingPreProvision",
 		Scopable: true,
-		Text:     "A command or script to execute before the resource :c-action:`provision` action. Errors interrupt the action.",
+		Text:     keywords.NewText(fs, "text/kw/blocking_pre_provision"),
 	}
 
 	KWBlockingPreRun = keywords.Keyword{
 		Option:   "blocking_pre_run",
 		Attr:     "BlockingPreRun",
 		Scopable: true,
-		Text:     "A command or script to execute before the resource :c-action:`run` action. Errors interrupt the action.",
+		Text:     keywords.NewText(fs, "text/kw/blocking_pre_run"),
 	}
 
 	KWBlockingPreStart = keywords.Keyword{
 		Option:   "blocking_pre_start",
 		Attr:     "BlockingPreStart",
 		Scopable: true,
-		Text:     "A command or script to execute before the resource :c-action:`start` action. Errors interrupt the action.",
+		Text:     keywords.NewText(fs, "text/kw/blocking_pre_start"),
 	}
 
 	KWBlockingPreStop = keywords.Keyword{
 		Option:   "blocking_pre_stop",
 		Attr:     "BlockingPreStop",
 		Scopable: true,
-		Text:     "A command or script to execute before the resource :c-action:`stop` action. Errors interrupt the action.",
+		Text:     keywords.NewText(fs, "text/kw/blocking_pre_stop"),
 	}
 
 	KWBlockingPreUnprovision = keywords.Keyword{
 		Option:   "blocking_pre_unprovision",
 		Attr:     "BlockingPreUnprovision",
 		Scopable: true,
-		Text:     "A command or script to execute before the resource :c-action:`unprovision` action. Errors interrupt the action.",
+		Text:     keywords.NewText(fs, "text/kw/blocking_pre_unprovision"),
 	}
 
 	KWDisable = keywords.Keyword{
@@ -81,7 +86,7 @@ var (
 		Attr:      "Disable",
 		Scopable:  true,
 		Converter: converters.Bool,
-		Text:      "A disabled resource will be ignored on service startup and shutdown. Its status is always reported ``n/a``.\n\nSet in DEFAULT, the whole service is disabled. A disabled service does not honor :c-action:`start` and :c-action:`stop` actions. These actions immediately return success.\n\n:cmd:`om <path> disable` only sets :kw:`DEFAULT.disable`. As resources disabled state is not changed, :cmd:`om <path> enable` does not enable disabled resources.",
+		Text:      keywords.NewText(fs, "text/kw/disable"),
 	}
 
 	KWEnableProvision = keywords.Keyword{
@@ -89,7 +94,7 @@ var (
 		Attr:      "EnableProvision",
 		Converter: converters.Bool,
 		Default:   "true",
-		Text:      "Set to false to skip the resource on provision and unprovision actions. Warning: Provision implies destructive operations like formating. Unprovision destroys service data.",
+		Text:      keywords.NewText(fs, "text/kw/provision"),
 	}
 
 	KWEnableUnprovision = keywords.Keyword{
@@ -97,7 +102,7 @@ var (
 		Attr:      "EnableUnprovision",
 		Converter: converters.Bool,
 		Default:   "true",
-		Text:      "Set to false to skip the resource on unprovision actions. Warning: Unprovision destroys service data.",
+		Text:      keywords.NewText(fs, "text/kw/unprovision"),
 	}
 
 	KWMonitor = keywords.Keyword{
@@ -105,7 +110,7 @@ var (
 		Attr:      "Monitor",
 		Scopable:  true,
 		Converter: converters.Bool,
-		Text:      "A down monitored resource will trigger a the monitor action (crash or reboot the node, freezestop or switch the service) if the monitor thinks the resource should be up and it all restart tries failed.",
+		Text:      keywords.NewText(fs, "text/kw/monitor"),
 	}
 
 	KWOptional = keywords.Keyword{
@@ -114,7 +119,7 @@ var (
 		Scopable:  true,
 		Converter: converters.Bool,
 		Inherit:   keywords.InheritHead2Leaf,
-		Text:      "Action failures on optional resources are logged but do not stop the action sequence. Also the optional resource status is not aggregated to the instance 'availstatus', but aggregated to the 'overallstatus'. Resource tagged :c-tag:`noaction` and sync resources are automatically considered optional. Useful for resources like dump filesystems for example.",
+		Text:      keywords.NewText(fs, "text/kw/optional"),
 	}
 
 	KWOptionalTrue = keywords.Keyword{
@@ -124,84 +129,84 @@ var (
 		Converter: converters.Bool,
 		Inherit:   keywords.InheritHead2Leaf,
 		Default:   "true",
-		Text:      "Action failures on optional resources are logged but do not stop the action sequence. Also the optional resource status is not aggregated to the instance 'availstatus', but aggregated to the 'overallstatus'. Resource tagged :c-tag:`noaction` and sync resources are automatically considered optional. Useful for resources like dump filesystems for example.",
+		Text:      keywords.NewText(fs, "text/kw/optional"),
 	}
 
 	KWPostProvision = keywords.Keyword{
 		Option:   "post_provision",
 		Attr:     "PostProvision",
 		Scopable: true,
-		Text:     "A command or script to execute after the resource :c-action:`provision` action. Errors do not interrupt the action.",
+		Text:     keywords.NewText(fs, "text/kw/post_provision"),
 	}
 
 	KWPostRun = keywords.Keyword{
 		Option:   "post_run",
 		Attr:     "PostRun",
 		Scopable: true,
-		Text:     "A command or script to execute after the resource :c-action:`run` action. Errors do not interrupt the action.",
+		Text:     keywords.NewText(fs, "text/kw/post_provision"),
 	}
 
 	KWPostStart = keywords.Keyword{
 		Option:   "post_start",
 		Attr:     "PostStart",
 		Scopable: true,
-		Text:     "A command or script to execute after the resource :c-action:`start` action. Errors do not interrupt the action.",
+		Text:     keywords.NewText(fs, "text/kw/post_provision"),
 	}
 
 	KWPostStop = keywords.Keyword{
 		Option:   "post_stop",
 		Attr:     "PostStop",
 		Scopable: true,
-		Text:     "A command or script to execute after the resource :c-action:`stop` action. Errors do not interrupt the action.",
+		Text:     keywords.NewText(fs, "text/kw/post_provision"),
 	}
 
 	KWPostUnprovision = keywords.Keyword{
 		Option:   "post_unprovision",
 		Attr:     "PostUnprovision",
 		Scopable: true,
-		Text:     "A command or script to execute after the resource :c-action:`unprovision` action. Errors do not interrupt the action.",
+		Text:     keywords.NewText(fs, "text/kw/post_provision"),
 	}
 
 	KWPreProvision = keywords.Keyword{
 		Option:   "pre_provision",
 		Attr:     "PreProvision",
 		Scopable: true,
-		Text:     "A command or script to execute before the resource :c-action:`provision` action. Errors do not interrupt the action.",
+		Text:     keywords.NewText(fs, "text/kw/post_provision"),
 	}
 
 	KWPreRun = keywords.Keyword{
 		Option:   "pre_run",
 		Attr:     "PreRun",
 		Scopable: true,
-		Text:     "A command or script to execute before the resource :c-action:`run` action. Errors do not interrupt the action.",
+		Text:     keywords.NewText(fs, "text/kw/post_provision"),
 	}
 
 	KWPreStart = keywords.Keyword{
 		Option:   "pre_start",
 		Attr:     "PreStart",
 		Scopable: true,
-		Text:     "A command or script to execute before the resource :c-action:`start` action. Errors do not interrupt the action.",
+		Text:     keywords.NewText(fs, "text/kw/post_provision"),
 	}
 
 	KWPreStop = keywords.Keyword{
 		Option:   "pre_stop",
 		Attr:     "PreStop",
 		Scopable: true,
-		Text:     "A command or script to execute before the resource :c-action:`stop` action. Errors do not interrupt the action.",
+		Text:     keywords.NewText(fs, "text/kw/post_provision"),
 	}
 
 	KWPreUnprovision = keywords.Keyword{
 		Option:   "pre_unprovision",
 		Attr:     "PreUnprovision",
 		Scopable: true,
-		Text:     "A command or script to execute before the resource :c-action:`unprovision` action. Errors do not interrupt the action.",
+		Text:     keywords.NewText(fs, "text/kw/post_provision"),
 	}
 
 	KWProvisionRequires = keywords.Keyword{
 		Option:  "provision_requires",
 		Attr:    "ProvisionRequires",
 		Example: "ip#0 fs#0(down,stdby down)",
-		Text:    "A whitespace-separated list of conditions to meet to accept doing a 'start' action. A condition is expressed as ``<rid>(<state>,...)``. If states are omitted, ``up,stdby up`` is used as the default expected states.",
+		Text:    keywords.NewText(fs, "text/kw/provision_requires"),
 	}
 
 	KWRestart = keywords.Keyword{
@@ -210,17 +215,7 @@ var (
 		Scopable:  true,
 		Converter: converters.Int,
 		Default:   "0",
-		Text: "The agent will try to restart a resource <n> times before falling back to the monitor action. A resource restart is triggered if:" +
-			"the resource is not disabled and its status is not up, " +
-			"and the node is not frozen, " +
-			"and the service instance is not frozen " +
-			"and its local expect is set to ``started``. " +
-			"If a resource has a restart set to a value greater than zero, its status is evaluated " +
-			"at the frequency defined by :kw:`DEFAULT.monitor_schedule` " +
-			"instead of the frequency defined by :kw:`DEFAULT.status_schedule`. " +
-			":kw:`restart_delay` defines the interval between two restarts. " +
-			"Standby resources have a particular value to ensure best effort to restart standby resources, " +
-			"default value is 2, and value lower than 2 are changed to 2.",
+		Text:      keywords.NewText(fs, "text/kw/restart"),
 	}
 
 	KWRestartDelay = keywords.Keyword{
@@ -229,29 +224,28 @@ var (
 		Scopable:  true,
 		Converter: converters.Duration,
 		Default:   "500ms",
-		Text: "Define minimum delay between two triggered restarts of a same resource (used when :kw:`restart`is defined). " +
-			"Default value is 0 (no delay).",
+		Text:      keywords.NewText(fs, "text/kw/post_provision"),
 	}
 
 	KWRunRequires = keywords.Keyword{
 		Option:  "run_requires",
 		Attr:    "RunRequires",
 		Example: "ip#0 fs#0(down,stdby down)",
-		Text:    "A whitespace-separated list of conditions to meet to accept doing a 'run' action. A condition is expressed as ``<rid>(<state>,...)``. If states are omitted, ``up,stdby up`` is used as the default expected states.",
+		Text:    keywords.NewText(fs, "text/kw/run_requires"),
 	}
 
 	KWSCSIPersistentReservationEnabled = keywords.Keyword{
 		Option:    "scsireserv",
 		Attr:      "SCSIPersistentReservation.Enabled",
 		Converter: converters.Bool,
-		Text:      "If set to ``true``, OpenSVC will try to acquire a type-5 (write exclusive, registrant only) scsi3 persistent reservation on every path to every disks held by this resource. Existing reservations are preempted to not block service start-up. If the start-up was not legitimate the data are still protected from being written over from both nodes. If set to ``false`` or not set, :kw:`scsireserv` can be activated on a per-resource basis.",
+		Text:      keywords.NewText(fs, "text/kw/scsireserv"),
 	}
 
 	KWSCSIPersistentReservationKey = keywords.Keyword{
 		Option:   "prkey",
 		Attr:     "SCSIPersistentReservation.Key",
 		Scopable: true,
-		Text:     "Defines a specific persistent reservation key for the resource. Takes priority over the service-level defined prkey and the node.conf specified prkey.",
+		Text:     keywords.NewText(fs, "text/kw/prkey"),
 	}
 
 	KWSCSIPersistentReservationNoPreemptAbort = keywords.Keyword{
@@ -259,7 +253,7 @@ var (
 		Attr:      "SCSIPersistentReservation.NoPreemptAbort",
 		Scopable:  true,
 		Converter: converters.Bool,
-		Text:      "If set to ``true``, OpenSVC will preempt scsi reservation with a preempt command instead of a preempt and and abort. Some scsi target implementations do not support this last mode (esx). If set to ``false`` or not set, :kw:`no_preempt_abort` can be activated on a per-resource basis.",
+		Text:      keywords.NewText(fs, "text/kw/no_preempt_abort"),
 	}
 
 	KWShared = keywords.Keyword{
@@ -267,7 +261,7 @@ var (
 		Attr:      "Shared",
 		Scopable:  true,
 		Converter: converters.Bool,
-		Text:      "Set to ``true`` to skip the resource on provision and unprovision actions if the action has already been done by a peer. Shared resources, like vg built on SAN disks must be provisioned once. All resources depending on a shared resource must also be flagged as shared.",
+		Text:      keywords.NewText(fs, "text/kw/shared"),
 	}
 
 	KWStandby = keywords.Keyword{
@@ -275,35 +269,35 @@ var (
 		Attr:      "Standby",
 		Scopable:  true,
 		Converter: converters.Bool,
-		Text:      "Always start the resource, even on standby instances. The daemon is responsible for starting standby resources. A resource can be set standby on a subset of nodes using keyword scoping.\n\nA typical use-case is sync'ed fs on non-shared disks: the remote fs must be mounted to not overflow the underlying fs.\n\n.. warning:: Don't set shared resources standby: fs on shared disks for example.",
+		Text:      keywords.NewText(fs, "text/kw/standby"),
 	}
 
 	KWStartRequires = keywords.Keyword{
 		Option:  "start_requires",
 		Attr:    "StartRequires",
 		Example: "ip#0 fs#0(down,stdby down)",
-		Text:    "A whitespace-separated list of conditions to meet to accept doing a 'start' action. A condition is expressed as ``<rid>(<state>,...)``. If states are omitted, ``up,stdby up`` is used as the default expected states.",
+		Text:    keywords.NewText(fs, "text/kw/start_requires"),
 	}
 
 	KWStopRequires = keywords.Keyword{
 		Option:  "stop_requires",
 		Attr:    "StopRequires",
 		Example: "ip#0 fs#0(down,stdby down)",
-		Text:    "A whitespace-separated list of conditions to meet to accept doing a 'stop' action. A condition is expressed as ``<rid>(<state>,...)``. If states are omitted, ``up,stdby up`` is used as the default expected states.",
+		Text:    keywords.NewText(fs, "text/kw/stop_requires"),
 	}
 
 	KWSubset = keywords.Keyword{
 		Option:   "subset",
 		Attr:     "Subset",
 		Scopable: true,
-		Text:     "Assign the resource to a specific subset.",
+		Text:     keywords.NewText(fs, "text/kw/post_provision"),
 	}
 
 	KWSyncRequires = keywords.Keyword{
 		Option:  "sync_requires",
 		Attr:    "SyncRequires",
 		Example: "ip#0 fs#0(down,stdby down)",
-		Text:    "A whitespace-separated list of conditions to meet to accept doing a 'sync' action. A condition is expressed as ``<rid>(<state>,...)``. If states are omitted, ``up,stdby up`` is used as the default expected states.",
+		Text:    keywords.NewText(fs, "text/kw/sync_requires"),
 	}
 
 	KWTags = keywords.Keyword{
@@ -311,14 +305,14 @@ var (
 		Attr:      "Tags",
 		Scopable:  true,
 		Converter: converters.Set,
-		Text:      "A list of tags. Arbitrary tags can be used to limit action scope to resources with a specific tag. Some tags can influence the driver behaviour. For example :c-tag:`noaction` avoids any state changing action from the driver and implies ``optional=true``, :c-tag:`nostatus` forces the status to n/a.",
+		Text:      keywords.NewText(fs, "text/kw/tags"),
 	}
 
 	KWUnprovisionRequires = keywords.Keyword{
 		Option:  "unprovision_requires",
 		Attr:    "UnprovisionRequires",
 		Example: "ip#0 fs#0(down,stdby down)",
-		Text:    "A whitespace-separated list of conditions to meet to accept doing a 'unprovision' action. A condition is expressed as ``<rid>(<state>,...)``. If states are omitted, ``up,stdby up`` is used as the default expected states.",
+		Text:    keywords.NewText(fs, "text/kw/unprovision_requires"),
 	}
 
 	SCSIPersistentReservationKeywords = []keywords.Keyword{