+++ drivers/resdiskcrypt/manifest.go
@@ -3,6 +3,8 @@
 package resdiskcrypt
 
 import (
+	"embed"
+
 	"github.com/opensvc/om3/core/driver"
 	"github.com/opensvc/om3/core/keywords"
 	"github.com/opensvc/om3/core/manifest"
@@ -11,6 +13,9 @@ import (
 )
 
 var (
+	//go:embed text
+	fs embed.FS
+
 	drvID = driver.NewID(driver.GroupDisk, "crypt")
 )
 
@@ -28,17 +33,17 @@ func (t T) Manifest() *manifest.T {
 			Option:      "name",
 			Attr:        "Name",
 			Scopable:    true,
-			Text:        "The basename of the exposed device.",
-			DefaultText: "The basename of the underlying device, suffixed with '-crypt'.",
 			Example:     "{fqdn}-crypt",
+			DefaultText: keywords.NewText(fs, "text/kw/name.default"),
+			Text:        keywords.NewText(fs, "text/kw/name"),
 		},
 		keywords.Keyword{
 			Option:   "dev",
 			Attr:     "Dev",
 			Scopable: true,
 			Required: true,
-			Text:     "The fullpath of the underlying block device.",
 			Example:  "/dev/{fqdn}/lv1",
+			Text:     keywords.NewText(fs, "text/kw/dev"),
 		},
 		keywords.Keyword{
 			Option:       "manage_passphrase",
@@ -47,13 +52,13 @@ func (t T) Manifest() *manifest.T {
 			Provisioning: true,
 			Converter:    converters.Bool,
 			Default:      "true",
-			Text:         "By default, on provision the driver allocates a new random passphrase (256 printable chars), and forgets it on unprovision. If set to false, require a passphrase to be already present in the sec object to provision, and don't remove it on unprovision.",
+			Text:         keywords.NewText(fs, "text/kw/manage_passphrase"),
 		},
 		keywords.Keyword{
 			Option:   "secret",
 			Attr:     "Secret",
 			Scopable: true,
-			Text:     "The name of the sec object hosting the crypt secrets. The sec object must be in the same namespace than the object defining the disk.crypt resource.",
+			Text:     keywords.NewText(fs, "text/kw/secret"),
 			Default:  "{name}",
 		},
 		keywords.Keyword{
@@ -61,7 +66,7 @@ func (t T) Manifest() *manifest.T {
 			Attr:         "FormatLabel",
 			Scopable:     true,
 			Provisioning: true,
-			Text:         "The label to set in the cryptsetup metadata writen on dev. A label helps admin understand the role of a device.",
+			Text:         keywords.NewText(fs, "text/kw/label"),
 			Default:      "{fqdn}",
 		},
 	)