+++ drivers/resipnetns/manifest.go
@@ -3,6 +3,8 @@
 package resipnetns
 
 import (
+	"embed"
+
 	"github.com/opensvc/om3/core/driver"
 	"github.com/opensvc/om3/core/keyop"
 	"github.com/opensvc/om3/core/keywords"
@@ -12,6 +14,9 @@ import (
 )
 
 var (
+	//go:embed text
+	fs embed.FS
+
 	drvID    = driver.NewID(driver.GroupIP, "netns")
 	altDrvID = driver.NewID(driver.GroupIP, "docker")
 )
@@ -32,7 +37,7 @@ func (t T) Manifest() *manifest.T {
 			Required: true,
 			Aliases:  []string{"container_rid"},
 			Example:  "container#0",
-			Text:     "The resource id of the container to plumb the ip into.",
+			Text:     keywords.NewText(fs, "text/kw/netns"),
 		},
 		keywords.Keyword{
 			Option:   "vlan_tag",
@@ -46,7 +51,7 @@ func (t T) Manifest() *manifest.T {
 				},
 			},
 			Example: "44",
-			Text:    "The VLAN tag the switch port will relay.",
+			Text:    keywords.NewText(fs, "text/kw/vlan_tag"),
 		},
 		keywords.Keyword{
 			Option:     "vlan_mode",
@@ -62,7 +67,7 @@ func (t T) Manifest() *manifest.T {
 			},
 			Default: "native-untagged",
 			Example: "access",
-			Text:    "The VLAN port mode.",
+			Text:    keywords.NewText(fs, "text/kw/vlan_mode"),
 		},
 		keywords.Keyword{
 			Option:     "mode",
@@ -71,22 +76,21 @@ func (t T) Manifest() *manifest.T {
 			Scopable:   true,
 			Default:    "bridge",
 			Example:    "access",
-			Text:       "The ip link mode. If ipdev is set to a bridge interface the mode defaults to bridge, else defaults to macvlan. ipvlan requires a 4.2+ kernel.",
+			Text:       keywords.NewText(fs, "text/kw/mode"),
 		},
 		keywords.Keyword{
-			Option:      "nsdev",
-			Attr:        "NSDev",
-			Scopable:    true,
-			DefaultText: "The first available eth<n>.",
-			Example:     "front",
-			Text:        "The interface name in the network namespace.",
+			Option:   "nsdev",
+			Attr:     "NSDev",
+			Scopable: true,
+			Example:  "front",
+			Text:     keywords.NewText(fs, "text/kw/nsdev"),
 		},
 		keywords.Keyword{
 			Option:   "macaddr",
 			Attr:     "MacAddr",
 			Scopable: true,
 			Example:  "ce:32:cc:ca:41:33",
-			Text:     "The hardware address to set on the interface in the network namespace.",
+			Text:     keywords.NewText(fs, "text/kw/macaddr"),
 		},
 		keywords.Keyword{
 			Option:    "del_net_route",
@@ -94,14 +98,14 @@ func (t T) Manifest() *manifest.T {
 			Converter: converters.Bool,
 			Scopable:  true,
 			Default:   "false",
-			Text:      "Some docker ip configuration requires dropping the network route autoconfigured when installing the ip address. In this case set this parameter to true, and also set the network parameter.",
+			Text:      keywords.NewText(fs, "text/kw/del_net_route"),
 		},
 		keywords.Keyword{
 			Option:   "ipname",
 			Attr:     "IpName",
 			Scopable: true,
 			Example:  "1.2.3.4",
-			Text:     "The DNS name or IP address of the ip resource. Can be different from one node to the other, in which case ``@nodename`` can be specified. This is most useful to specify a different ip when the service starts in DRP mode, where subnets are likely to be different than those of the production datacenter. With the amazon driver, the special ``<allocate>`` value tells the provisioner to assign a new private address.",
+			Text:     keywords.NewText(fs, "text/kw/ipname"),
 		},
 		keywords.Keyword{
 			Option:   "ipdev",
@@ -109,20 +113,20 @@ func (t T) Manifest() *manifest.T {
 			Scopable: true,
 			Example:  "eth0",
 			Required: true,
-			Text:     "The interface name over which OpenSVC will try to stack the service ip. Can be different from one node to the other, in which case the ``@nodename`` can be specified. If the value is expressed as '<intf>:<alias>, the stacked interface index is forced to <alias> instead of the lowest free integer. If the value is expressed as <name>@<intf>, a macvtap interface named <name> is created and attached to <intf>.",
+			Text:     keywords.NewText(fs, "text/kw/ipdev"),
 		},
 		keywords.Keyword{
 			Option:   "netmask",
 			Attr:     "Netmask",
 			Scopable: true,
 			Example:  "24",
-			Text:     "If an ip is already plumbed on the root interface (in which case the netmask is deduced from this ip). Mandatory if the interface is dedicated to the service (dummy interface are likely to be in this case). The format is either dotted or octal for IPv4, ex: 255.255.252.0 or 22, and octal for IPv6, ex: 64.",
+			Text:     keywords.NewText(fs, "text/kw/netmask"),
 		},
 		keywords.Keyword{
 			Option:       "gateway",
 			Attr:         "Gateway",
 			Scopable:     true,
-			Text:         "A zone ip provisioning parameter used in the sysidcfg formatting. The format is decimal for IPv4, ex: 255.255.252.0, and octal for IPv6, ex: 64.",
+			Text:         keywords.NewText(fs, "text/kw/gateway"),
 			Provisioning: true,
 		},
 		keywords.Keyword{
@@ -130,37 +134,15 @@ func (t T) Manifest() *manifest.T {
 			Attr:      "WaitDNS",
 			Scopable:  true,
 			Converter: converters.Bool,
-			Text:      "Wait for the cluster DNS records associated to the resource to appear after a resource start and before the next resource can be started. This can be used for apps or containers that require the ip or ip name to be resolvable to provision or execute properly.",
-		},
-		keywords.Keyword{
-			Option:   "dns_name_suffix",
-			Attr:     "DNSNameSuffix",
-			Scopable: true,
-			Text:     "Add the value as a suffix to the DNS record name. The record created is thus formatted as ``<name>-<dns_name_suffix>.<app>.<managed zone>``.",
-		},
-		keywords.Keyword{
-			Option:       "provisioner",
-			Attr:         "Provisioner",
-			Scopable:     true,
-			Candidates:   []string{"collector", ""},
-			Example:      "collector",
-			Text:         "The IPAM driver to use to provision the ip.",
-			Provisioning: true,
+			Text:      keywords.NewText(fs, "text/kw/wait_dns"),
 		},
 		keywords.Keyword{
 			Option:       "network",
 			Attr:         "Network",
 			Scopable:     true,
 			Example:      "10.0.0.0/16",
-			Text:         "The network, in dotted notation, from where the ip provisioner allocates. Also used by the docker ip driver to delete the network route if :kw:`del_net_route` is set to ``true``.",
+			Text:         keywords.NewText(fs, "text/kw/network"),
 			Provisioning: true,
-		},
-		keywords.Keyword{
-			Option:    "dns_update",
-			Attr:      "DNSUpdate",
-			Scopable:  true,
-			Converter: converters.Bool,
-			Text:      "Setting this parameter triggers a DNS update. The record created is formatted as ``<name>.<app>.<managed zone>``, unless dns_record_name is specified.",
 		},
 		keywords.Keyword{
 			Option:    "check_carrier",
@@ -168,7 +150,7 @@ func (t T) Manifest() *manifest.T {
 			Scopable:  true,
 			Default:   "true",
 			Converter: converters.Bool,
-			Text:      "Activate the link carrier check. Set to false if ipdev is a backend bridge or switch.",
+			Text:      keywords.NewText(fs, "text/kw/check_carrier"),
 		},
 		keywords.Keyword{
 			Option:    "alias",
@@ -176,7 +158,7 @@ func (t T) Manifest() *manifest.T {
 			Scopable:  true,
 			Default:   "true",
 			Converter: converters.Bool,
-			Text:      "Use ip aliasing. Modern ip stack support multiple ip/mask per interface, so :kw:`alias` should be set to false when possible.",
+			Text:      keywords.NewText(fs, "text/kw/alias"),
 		},
 		keywords.Keyword{
 			Option:    "expose",
@@ -184,7 +166,7 @@ func (t T) Manifest() *manifest.T {
 			Scopable:  true,
 			Converter: converters.List,
 			Example:   "443/tcp:8443 53/udp",
-			Text:      "A whitespace-separated list of ``<port>/<protocol>[:<host port>]`` describing socket services that mandate a SRV exposition. With <host_port> set, the ip.cni driver configures port mappings too.",
+			Text:      keywords.NewText(fs, "text/kw/expose"),
 		},
 	)
 	return m