+++ drivers/resiproute/manifest.go
@@ -1,12 +1,17 @@
 package resiproute
 
 import (
+	"embed"
+
 	"github.com/opensvc/om3/core/driver"
 	"github.com/opensvc/om3/core/keywords"
 	"github.com/opensvc/om3/core/manifest"
 )
 
 var (
+	//go:embed text
+	fs embed.FS
+
 	drvID = driver.NewID(driver.GroupIP, "route")
 )
 
@@ -23,7 +28,7 @@ func (t T) Manifest() *manifest.T {
 			Attr:     "NetNS",
 			Scopable: true,
 			Required: true,
-			Text:     "the resource id of the container to plumb the ip into.",
+			Text:     keywords.NewText(fs, "text/kw/netns"),
 			Example:  "container#0",
 		},
 		keywords.Keyword{
@@ -31,7 +36,7 @@ func (t T) Manifest() *manifest.T {
 			Attr:     "Gateway",
 			Scopable: true,
 			Required: true,
-			Text:     "the gateway ip address.",
+			Text:     keywords.NewText(fs, "text/kw/gateway"),
 			Example:  "1.2.3.4",
 		},
 		keywords.Keyword{
@@ -39,7 +44,7 @@ func (t T) Manifest() *manifest.T {
 			Attr:     "To",
 			Scopable: true,
 			Required: true,
-			Text:     "The route destination node.",
+			Text:     keywords.NewText(fs, "text/kw/to"),
 			Example:  "192.168.100.0/24",
 		},
 		keywords.Keyword{
@@ -47,8 +52,8 @@ func (t T) Manifest() *manifest.T {
 			Attr:        "Dev",
 			Scopable:    true,
 			Required:    false,
-			DefaultText: "Any first dev with an addr in the same network than the gateway.",
-			Text:        "The network link to add the route on.",
+			DefaultText: keywords.NewText(fs, "text/kw/dev.default"),
+			Text:        keywords.NewText(fs, "text/kw/dev"),
 			Example:     "eth1",
 		},
 	)